@@ -103,16 +120,35 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.1</version>
+                <version>3.13.0</version>
                 <configuration>
-                    <source>1.7</source>
-                    <target>1.7</target>
+                    <source>21</source>
+                    <target>21</target>
+                </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <version>2.22.2</version>
+                <configuration>
+                    <forkCount>1</forkCount>
+                    <reuseForks>true</reuseForks>
+                    <!--					<argLine>-Djdk.attach.allowAttachSelf</argLine>-->
+                    <argLine>
+                        -javaagent:${settings.localRepository}/org/jmockit/jmockit/1.49/jmockit-1.49.jar
+                    </argLine>
+                    <useSystemClassLoader>true</useSystemClassLoader>
+                    <!--					<forkMode>once</forkMode>-->
+
+                    <excludes>
+                        <exclude>**/integration/JobControllerTest.java</exclude>
+                    </excludes>
                 </configuration>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-enforcer-plugin</artifactId>
-                <version>1.2</version>
+                <version>3.5.0</version>
                 <executions>
                     <execution>
                         <id>enforce</id>
